Punycode domains works!
Ugly version for proof of concept. To avoid having to make these calculations
time and time again, the hostname ought to be stored in one unicode representation
(used whenever possible in out-going communication, for the sake of visual appearance)
and one in punycode version.

In-coming webfinger requests will request with either domain name (unicode or punycode),
and some even request twice: one with each version of the domain name.