ActivityPub: Use classes to match hashtags
Hashtags are defined to consist of only alphanumeric characters plus
underscore, with digits only not allowed. We can match that in a regular
expression when making use of the predefined character classes.

This requires LC_CTYPE to use an UTF-8 locale in order to be correct for
the entire Unicode range, not just for the ASCII range, therefore set
LC_CTYPE to whatever is the environment for now.

Relates:	grunfink/snac2#199