xs_tolower_i: Fix for characters >= 0x80
The argument for tolower() must be the value of an *unsigned* character
converted to `int`, a negative value is invalid. Some implementations,
e.g. on NetBSD, will just return `\0` for every negative value. Add an
explicit cast to `unsigned char` to avoid this issue when the default
`char` is signed.

Note this function is used for processing tags, which are allowed to
contain characters outside ASCII. For correct processing, alphabetic
characters outside ASCII should also be lowercased, which is not
possible with tolower(). A fully correct solution should therefore
convert tags to a wide character string and use towlower() instead.